[BE] feat: 중도퇴장 기능 추가 및 시작시간 도메인 변경

+++ BE/project/src/main/java/com/ssafy/project/repository/BookParticipationRecordRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.project.repository;
 import com.ssafy.project.domain.record.BookParticipationRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookParticipationRecordRepository extends JpaRepository<BookParticipationRecord, Integer> {
+public interface BookParticipationRecordRepository extends JpaRepository<BookParticipationRecord, Long> {
 }
